Patch to fix problem noted by "Ing. Jozef Sakalos" <jsakalos@ba.success.sk> with comparisons between signed an unsigned ints.
CVS patchset: 2131
CVS date: 2003/07/22 09:51:35

+++ lib/plugins/stonith/baytech.c
@@ -153,7 +153,7 @@ struct BayTech {
 
 struct BayTechModelInfo {
 	const char *	type;		/* Baytech model info */
-	int		socklen;	/* Length of socket name string */
+	size_t		socklen;	/* Length of socket name string */
 	struct Etoken *	expect;		/* Expect string before outlet list */
 };
